Move cached embedding devices and dtype for onnx export consistency

+++ nemo/collections/multimodal/modules/stable_diffusion/diffusionmodules/util.py
@@ -24,13 +24,12 @@ thanks!
 '''
 
 import math
-from inspect import isfunction
-
 import numpy as np
 import torch
 import torch.nn as nn
 from apex.contrib.group_norm import GroupNorm
 from einops import repeat
+from inspect import isfunction
 from torch._dynamo import disable
 from torch.cuda.amp import custom_bwd, custom_fwd
 
@@ -207,7 +206,7 @@ def timestep_embedding(timesteps, dim, max_period=10000, repeat_only=False, cach
     if not repeat_only:
         if cached_embedding is not None:
             # using cached embedding and lookup in the cache
-            embedding = cached_embedding[timesteps, :]
+            embedding = cached_embedding[timesteps.to(dtype=torch.int), :]
         else:
             half = dim // 2
             idx = get_idx(half, timesteps.device)
